+++ src/api/api.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { Stats } from './dto/stats.dto'
-import { EsdtToken } from './dto/esdtToken.dto';
+import { EsdtToken } from './dto/esdtToken';
 import { createJsonObject } from './utils/readJsonFile';
 import { NftCollection } from './dto/nftCollection.dto';
